chore: Refactors individual_user_template
