+++ rust-agent/src/individual_user_template/mod.rs
@@ -1,5 +1,6 @@
 #![allow(dead_code, unused_imports)]
 mod individual_user_template_ffi;
+mod individual_user_template_helper;
 use candid::{self, CandidType, Decode, Deserialize, Encode, Principal};
 use ic_agent::{export::{reqwest::Error, PrincipalError}, identity::{self, DelegatedIdentity}, Agent};
 use k256::elliptic_curve::JwkEcKey;
